requirements.txt
Flask==2.3.3
scikit-learn==1.4.1
numpy==1.26.4
cryptography==42.0.5
pycryptodome==3.20.0
requests==2.31.0
